update to java 1.8
	remove unecessary initialisations
	update to variadicts arguments
	simplify a lot of conditionals tests
	remove unecessary try/catch and throws of Exceptions
	...

+++ bluecove-bluez/org/bluecove/socket/LocalServerSocket.java
@@ -123,6 +123,6 @@ public class LocalServerSocket extends ServerSocket {
             throw new SocketException("Socket is already closed");
         }
         Object value = impl.getOption(LocalSocketOptions.SO_PASSCRED);
-        return value instanceof Integer && ((Integer) value).intValue() > 0;
+        return value instanceof Integer && (Integer) value > 0;
     }
 }
